[standalone]: Move SDK auth-config types to embedding-sdk-shared
