+++ frontend/src/embedding-sdk-bundle/types/auth-config.ts
@@ -1,95 +1,4 @@
-import type { MetabaseFetchRequestTokenFn } from "metabase/embedding-sdk/types/refresh-token";
-
-/**
- * @inline
- */
-type BaseMetabaseAuthConfig = {
-  metabaseInstanceUrl: string;
-};
-
-/**
- * @category MetabaseProvider
- */
-export type MetabaseAuthConfigWithJwt = BaseMetabaseAuthConfig & {
-  /**
-   * Which authentication method to use.
-   * If both SAML and JWT are enabled at the same time,
-   * it defaults to SAML unless the preferredAuthMethod is specified.
-   */
-  preferredAuthMethod?: "jwt";
-
-  /**
-   * Uri of the jwt provider. If provided the sdk will use jwt and will skip the first `/auth/sso` discovery request.
-   */
-  jwtProviderUri?: string;
-
-  /**
-   * Specifies a function to fetch the refresh token.
-   * The refresh token should be in the format of {@link UserBackendJwtResponse}
-   */
-  fetchRequestToken?: MetabaseFetchRequestTokenFn;
-
-  isGuest?: false;
-  apiKey?: never;
-};
-
-/**
- * @category MetabaseProvider
- */
-export type MetabaseAuthConfigWithSaml = BaseMetabaseAuthConfig & {
-  /**
-   * Which authentication method to use.
-   * If both SAML and JWT are enabled at the same time,
-   * it defaults to SAML unless the preferredAuthMethod is specified.
-   */
-  preferredAuthMethod?: "saml";
-  isGuest?: false;
-  apiKey?: never;
-  fetchRequestToken?: never;
-};
-
-/**
- * @category MetabaseProvider
- */
-export type MetabaseAuthConfigWithApiKey = BaseMetabaseAuthConfig & {
-  apiKey: string;
-  isGuest?: false;
-  preferredAuthMethod?: never;
-  fetchRequestToken?: never;
-};
-
-/**
- * @category MetabaseProvider
- */
-export type MetabaseIsGuestAuthConfig = BaseMetabaseAuthConfig & {
-  /**
-   * Defines if SDK should work in a Guest Embed mode
-   */
-  isGuest: true;
-
-  /**
-   * URL endpoint for fetching and refreshing guest embed JWT tokens (iframe only, not applicable for SDK's guest mode).
-   * Supports both token refresh on expiry and initial token fetch when no static token is provided.
-   * In both cases, this works with guest embed components (metabase-dashboard and metabase-question).
-   * The endpoint should return { jwt: string } with the new token.
-   */
-  guestEmbedProviderUri?: string;
-
-  apiKey?: never;
-  preferredAuthMethod?: never;
-  fetchRequestToken?: never;
-};
-
-/**
- * @category MetabaseProvider
- */
-export type MetabaseAuthConfig =
-  | MetabaseAuthConfigWithApiKey
-  | MetabaseAuthConfigWithJwt
-  | MetabaseAuthConfigWithSaml
-  | MetabaseIsGuestAuthConfig;
-
-export type MetabaseAuthMethod = Exclude<
-  MetabaseAuthConfig["preferredAuthMethod"],
-  undefined
->;
+// Auth config types moved to embedding-sdk-shared so non-app consumers can import
+// them without crossing module boundaries. Re-exported here so existing
+// `embedding-sdk-bundle/types/auth-config` imports keep working.
+export type * from "embedding-sdk-shared/types/auth-config";
