+++ frontend/src/embedding-sdk-shared/lib/define-metabase-auth-config.ts
@@ -1,4 +1,4 @@
-import type { MetabaseAuthConfig } from "embedding-sdk-bundle/types";
+import type { MetabaseAuthConfig } from "embedding-sdk-shared/types/auth-config";
 
 export const defineMetabaseAuthConfig = (
   config: MetabaseAuthConfig,
